Refactor: 파일 구조 변경, ScrollTop 분리
