+++ .vscode/conponent.code-snippets
@@ -5,7 +5,6 @@
 			"/** @jsxImportSource @emotion/react */",
 			"import { css } from '@emotion/react';",
 			"import { Box } from '@mui/material';",
-			"import React from 'react';",
 			"",
 			"type ${1}Props = {",
 			"  ",
