+++ components/home/Project.tsx
@@ -1,6 +1,6 @@
 /** @jsxImportSource @emotion/react */
 import { css } from '@emotion/react';
-import { Box, Card, Grid, styled } from '@mui/material';
+import { Box, Grid, styled } from '@mui/material';
 import { ResponsiveLayout, Text } from '../common';
 import HomeHead from './HomeHead';
 import Slider from 'react-slick';
@@ -8,7 +8,6 @@ import ArrowForwardIosIcon from '@mui/icons-material/ArrowForwardIos';
 import ArrowBackIosIcon from '@mui/icons-material/ArrowBackIos';
 import { testProject1, testProject2 } from '../../types';
 import { ProjectCard, ProjectDetailCard } from '../Card';
-import ScrollTop from '../ScrollTop';
 
 type ProjectProps = {};
 
@@ -67,8 +66,6 @@ const Project = ({}: ProjectProps) => {
           </Grid>
         </Section>
       </ResponsiveLayout>
-
-      <ScrollTop />
     </Box>
   );
 };
