+++ layouts/MainLayout.tsx
@@ -1,7 +1,8 @@
 import { Box } from '@mui/material';
 import React from 'react';
-import Footer from '../components/Footer';
-import MainNav from '../components/MainNav';
+import Footer from './Footer';
+import MainNav from './MainNav';
+import ScrollTop from './ScrollTop';
 
 type MainLayoutProps = {
   children: React.ReactChild;
@@ -12,6 +13,7 @@ const MainLayout = ({ children }: MainLayoutProps) => {
     <>
       <MainNav />
       {children}
+      <ScrollTop />
       <Footer />
     </>
   );
