+++ pages/blog/[id].tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Blog from '../../components/home/Blog';
-import Main from '../../components/Main';
-import MainLayout from '../../layouts/MinLayout';
+import Main from '../../components/home/Main';
+import MainLayout from '../../layouts/MainLayout';
 
 type BlogPageProps = {};
