+++ pages/member.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Member from '../components/home/Member';
-import MainLayout from '../layouts/MinLayout';
+import MainLayout from '../layouts/MainLayout';
 
 const MemberPage = () => {
   return (
