+++ pages/project.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Project from '../components/home/Project';
-import MainLayout from '../layouts/MinLayout';
+import MainLayout from '../layouts/MainLayout';
 
 const ProjectPage = () => {
   return (
